requirements.txt
DrissionPage==4.1.0.9
requests
Flask==2.3.3
SQLAlchemy==2.0.23
pymysql==1.1.0
flask_sqlalchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.0